+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Release History
 
+- Unreleased
+    - New SimCapiMessage Types:
+        - REGISTER_LOCAL_DATA_CHANGE_LISTENER,
+        - REGISTERED_LOCAL_DATA_CHANGED
+       *        New public transporter method: registerLocalDataListener
 - 1.2.1 (05 Dec 2016)
     - Some clean up and making the project Jenkins ready.
 - 1.2.0 (21 Oct 2016)